+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## NEXT
+
+# 4.5.10
 1. [dart]添加HDR/HEIC/HEIF/TIFF/WBMP/WEBP等图片格式的测试案例
 2. [ohos]解决透明弹窗页面切后台或被Native页面覆盖后出现的假死问题